♻️ Convert Features/Home to KMP
Converting the feature module to KMP and moving all the possible parts
to be multiplatform. The things that can't be ported now are disabled
for future changes.
