+++ features/task/src/commonMain/kotlin/com/escodro/task/mapper/CategoryMapper.kt
@@ -1,0 +1,14 @@
+package com.escodro.task.mapper
+
+import com.escodro.categoryapi.model.Category as ViewCategory
+import com.escodro.domain.model.Category as DomainCategory
+
+internal class CategoryMapper {
+
+    fun toView(domainCategory: DomainCategory): ViewCategory =
+        ViewCategory(
+            id = domainCategory.id,
+            name = domainCategory.name,
+            color = 0, // TODO convert to color in KMP
+        )
+}
