+++ features/task/src/commonMain/kotlin/com/escodro/task/presentation/detail/main/TaskDetailState.kt
@@ -1,0 +1,12 @@
+package com.escodro.task.presentation.detail.main
+
+import com.escodro.task.model.Task
+
+internal sealed class TaskDetailState {
+
+    data object Loading : TaskDetailState()
+
+    data object Error : TaskDetailState()
+
+    data class Loaded(val task: Task) : TaskDetailState()
+}
