--- features/task/src/main/java/com/escodro/task/mapper/CategoryMapper.kt
@@ -1,15 +1,0 @@
-package com.escodro.task.mapper
-
-import android.graphics.Color
-import com.escodro.categoryapi.model.Category as ViewCategory
-import com.escodro.domain.model.Category as DomainCategory
-
-internal class CategoryMapper {
-
-    fun toView(domainCategory: DomainCategory): ViewCategory =
-        ViewCategory(
-            id = domainCategory.id,
-            name = domainCategory.name,
-            color = Color.parseColor(domainCategory.color),
-        )
-}
