--- features/task/src/main/java/com/escodro/task/presentation/detail/alarm/AlarmSelectionState.kt
@@ -1,44 +1,0 @@
-package com.escodro.task.presentation.detail.alarm
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import com.escodro.task.model.AlarmInterval
-import java.util.Calendar
-
-/**
- * State holder for the [AlarmSelection] composable.
- */
-internal class AlarmSelectionState(calendar: Calendar?, alarmInterval: AlarmInterval?) {
-
-    /**
-     * The alarm date, if set.
-     */
-    var date by mutableStateOf(calendar)
-
-    /**
-     * The alarm data, if set.
-     */
-    var alarmInterval by mutableStateOf(alarmInterval)
-
-    /**
-     * The Exact Alarm permission dialog visibility state.
-     */
-    var showExactAlarmDialog by mutableStateOf(false)
-
-    /**
-     * The Notification permission dialog visibility state.
-     */
-    var showNotificationDialog by mutableStateOf(false)
-
-    /**
-     * The Notification Rationale dialog visibility state.
-     */
-    var showRationaleDialog by mutableStateOf(false)
-}
-
-@Composable
-internal fun rememberAlarmSelectionState(calendar: Calendar?, alarmInterval: AlarmInterval?) =
-    remember { AlarmSelectionState(calendar, alarmInterval) }
