--- features/task/src/main/res/values/strings.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="task_cd_add_task">Add task</string>
-    <string name="task_list_header_empty">Wow! All tasks are completed!</string>
-    <string name="task_list_cd_empty_list">That\'s all clear</string>
-    <string name="task_list_header_error">Ops… That\'s embarrassing…</string>
-    <string name="task_list_cd_error">Something went wrong</string>
-    <string name="task_detail_header_error">This task is no longer available</string>
-    <string name="task_detail_cd_error">Task not found</string>
-
-    <string name="task_detail_cd_icon_category">Categories</string>
-    <string name="task_detail_cd_icon_description">Description</string>
-    <string name="task_detail_cd_icon_alarm">Alarm</string>
-    <string name="task_detail_cd_icon_remove_alarm">Remove alarm</string>
-    <string name="task_detail_cd_icon_repeat_alarm">Repeating alarm</string>
-
-    <string name="task_detail_category_empty_list">No categories</string>
-
-    <string name="task_detail_alarm_no_alarm">No alarm</string>
-
-    <string name="task_add_save">Add</string>
-    <string name="task_add_label">Task</string>
-
-    <string name="task_snackbar_message_complete">\"%s\" completed</string>
-    <string name="task_snackbar_button_undo">Undo</string>
-
-    <string name="task_alarm_permission_dialog_title">Alarm and reminders</string>
-    <string name="task_alarm_permission_dialog_text">In order to alarms work, Alkaa needs the permission to set alarms and reminders. If this permission is not granted, the existing alarms will not work as well.</string>
-    <string name="task_alarm_permission_dialog_confirm">Grant permission</string>
-    <string name="task_alarm_permission_dialog_cancel">Not now</string>
-
-    <string name="task_notification_permission_dialog_title">Notifications</string>
-    <string name="task_notification_permission_dialog_text">In order to notify the alarms, Alkaa needs the permission to launch notifications. If this permission is not granted, the existing alarms will not work as well.</string>
-    <string name="task_notification_permission_dialog_confirm">Grant permission</string>
-    <string name="task_notification_permission_dialog_cancel">Not now</string>
-
-    <string name="task_notification_rationale_dialog_title">Notifications</string>
-    <string name="task_notification_rationale_dialog_text">It\'s not possible to set alarms without notification permission. If you want to use this feature, please go to app settings and grant it manually.</string>
-    <string name="task_notification_rationale_dialog_confirm">Take me there</string>
-    <string name="task_notification_rationale_dialog_cancel">Ignore</string>
-</resources>
